Added --with-gcc-flags and fixed code to compile cleanly with the flags. Original: 54d7fd3715f5851b91fc84e57d5cc4412f0ed16c

+++ lal/packages/utilities/src/Random.c
@@ -67,7 +67,7 @@ LALCreateRandomParams (
     (*params)->i = BasicRandom ((*params)->i);
   }
 
-  for (n = 0; n < sizeof((*params)->v)/sizeof(INT4); ++n)
+  for (n = 0; n < (int)(sizeof((*params)->v)/sizeof(INT4)); ++n)
   {
     (*params)->v[n] = (*params)->i = BasicRandom ((*params)->i);
   }
